[guest] Add #[codec] and #[with_serde] attributes
[codec] stops the user from having to list out all the derives they
need. You can of course not use it if you want. I couldn't do it as a
derive macro unfortunately since rust doesn't seem to make it easy to
call other derives from your derive.

[with_serde] allows you to specify function parameters that don't have
bincode implemented for them but do have a serde implementation we can
use instead.

Also renamed #[carol] to #[machine] since I felt that described what is
was better.
